Use a callback instead of the session for redirect
A callback is more versatile than a simple session value. By
using a "client callback" that is called only on the client we
can ensure that the callback is executed before the method call
ends. By only setting the captured variable 'id', we ensure that
meteor method call can be finished before we redirect.

The empty callbacks for the minutes.insert/update calls are
required, otherwise meteor would invoke the client callback only
after the server is finished.