Delete the local jquery copy of datetimepicker
npm downloads a local copy of jquery for the eonasdan/bootstrap-datetimepicker
module. The module then adds its functions to the wrong jquery instance. To
ensure that the functions are added to the correct instance we need to
load the same jquery module that meteor loads and that is the one in
node_modules/jquery. This is accomplished by deleting the local copy of
the datetimepicker.

Add momentjs