Try to make builds more stable
This introduces a setup script that runs the meteor end2end server and
kills it as soon as it has started and then exits itself. The advantage
of this is that we now have a install step that downloads all packages.
Since it is not started in the background Travis can wait for the initial
meteor setup before proceeding with the tests.