Fix transition from LDAP to following E2E tests
LDAP tests need to log out the current user before each test. If the last LDAP test happens to be one that ends in a logged out user the following tests might fail if they do not login the user explicitly. To prevent that we log in the default user after the LDAP tests are done.