Clear flash message on route change
When an error occurs and the user is informed via the flashMessage.js/errorMessage.html helpers in client/helpers, it can happen that the error will also be shown on another page if the user goes to that page quick enough.

Example: The user tries to login, an error occurs and he is notified via the errorMessage template. If he now logs in and opens the Add Meeting Series form within 5 seconds,  the error will still be displayed.

This change fixes this issue by removing all pending messages before  the route changes.