Simple filter for Items list
Created a simple Filter UI-Component providing a input field to define a query to filter elements. This component will be used within the Items tab to filter the list of info/action items. The query string will be evaluated by the TopicFilter-Class which applies the query to a given array of topics. Added unit tests to test the TopicFiler-Class.