Improved UX for finalizing minutes
During the finalize process a flash message will be displayed to show there is something going on. This might be helpful if the finalize process takes a while. After the process was finished the progress message will be replaced by a success message. So in case of a fast finalize-process the message will not be faded out too fast.

Furthermore the FlashMessage module was improved: if a component displays a flash message short after another message was displayed the first one will be shown at least 1.2 seconds. All following messages will be queued and displayed with a delay of 1.2 seconds.