Properly implement user permissions in minutes edit template
There are two new template helpers that help with the permission checks: authenticating and canShow.

authenticating() returns true if the user is currently being logged in and while we are waiting for the minute subscriptions. Once the minutes subscription is available we use canShow to check if the user is authorized to see the meeting minute. If he is, we render the original minutesedit template. If he is not allowed to see it we redirect him to '/'.

This causes the Template.onRendered() handler to be useless for our purposes. It can now be fired before blaze even begins to render the meeting minutes. But we need to know when certain HTML elements arrive on the page so we can initialize both the datetime picker and the topic panels properly.

To achieve this we (for now) use a dummy template helper that is triggered once the minutes is displayed and uses arrive.js to trigger the initialization once the required elements are entering the document.