Fix: The demo account was always created.
Even if settings.json setting was "false" we created a demo account on startup.

**Lesson learned:**
If one provides a function for a boolean predicate.
And one forgets to use braces on usage like so:
    if (GlobalSettings.createDemoAccount) ...
then this will evaluate to "true", as the test just checks, of the object has "something" with this name. And yes: it has a function with this name... So this evaluates to true - regadless of what the function call would return.

So right usage is clearly:
    if (GlobalSettings.createDemoAccount()) ...