E2E: Fix non-disabled "create new minutes" button
With switch to new meteor version - for whatever reason (TM) - the old way of disabling a <a> button via "disabled=true/false" way does not work anymore. The button is *always* enabled. Our E2E test reported this correctly. 
As a work around we now disable the <a> button via a parent <fieldset> tag. If this tag has a disabled attribute (regadless of its value) all embedded controlls are disabled. This is also the proposed way of the bootstrap material design framework.