Fix E2E-Test 'can not add minutes if unfinalized minutes exist'
Unfortunately checking if the button is disabled or is not click-able does not work. Instead we now verify that neither the current url nor the amount of minutes does not change. This does not verify that the button is not in a disabled state. Instead it verifies that nothing happens if the button will be pressed.