Add new detail element directly after closing the new-item-dialog
To realise that, I abused the magic of session variables. If a new item was created within the dialog a special session variable will be set with the new id as its value. The topicInfoItem-Template has a special template helper which will be executed whenever this session variable changes. If the variable contains the item id of the template it takes care about adding the new detail item.