Fix: docker deployment did not shutdown embedded mongodb resulting in "unclean shutdown" error messages
Lessons Learned:
1.)
Dockerfile needs to launch our 4minitz.sh script with the ["/4minitz.sh"] syntax. This ensures that the /bin/bash process gets the PID==1. And only PID==1 will receive the SIGINT signal on "Ctrl+C" stop.

2.)
Further we need a "trap" signalhandler() in our warpper 4minitz.sh script.

3.) 
For "docker stop" to work, we additionally needed to launch our node process in the background with "&" and a following "wait". This ensures that the script and not node.js gets the SIGINT from "docker stop".