Updated unit tests for minutes.upsertTopic
It is not necessary anymore to check if the topics array of the minutes has changed. We call a meteor method to explicitly add the topic which updates the database directly. This invokes the meteor magic to update the clients database so we don't have to worry about updating the topics array ourselves.