Improved arrangement of topic-controls in minutesedit-view
We differentiate between two situations: read-only-mode and screen-size. On screen with more than 768px width we display all three buttons in one row. On screens with less than 970px width we hide the word topics (except for the add-topics-button. There we hide this word only on devices less than 768px).

In the read-only-mode we do not have the add-topic-button, so we have to arrange only three controls: collapse, expand and hide-closed-topics. On small devices with less than 768px we arrange them in two rows with collapse/expand in the first and the hide-closed-checkbox in the second row.

In the edit-mode we have to arrange all four controls. Again, on small devices with less than 768px we arrange them in two rows with the add-topic button and the hide-closed-checkbox in the first and the collapse/expand controls in the second row.

The edit-mode looks weird on devices with less than 333px and the read-only-mode on devices with less than 280px. But I think this is ok.