Fixed broken end2end-test
Due to our responsive view we have two hide-close-topics controls on our page. One is visible on large devices and the other is visible on smaller devices. Since our end2end tests run only on a large screen (otherwise the menu would not work) I removed the id of the control visible on small devices. Now there is only one control which should be clicked for the test case.