Enh: updateChecks to inform adminsabout new released versions
Master (reports his version number) and is intended to run on 4minitz.com. Slave is every installed client that checks every 8 hours if there is a newer version available. If the found new version was not already reported, feature will log to console and send hint email to all admins.
* Master is per default off. Must be switched on via settings.json: checkUpdateMaster: true
* Slave ist per default on. Can be turned off via settings.json checkUpdate: false