Removed code duplication by providing a label-extractor service, plus extended algorithm for label detection
We had the detection of labels in a string in both topics and info item classes. This is now moved into a separate module the labelExtractor which recognizes labels within a string if they start with a # (this is always necessary) and they either are already known labels of this series (string-matching) or they match our regular expression (starting with a # then at least one character but no whitespaces).