Fix: Recurring attribute must be removed on delete. Fixes #317
Topics inherited from old meetings cannot be deleted. Instead they are "closed" after security question to user, so they do not show up in next meeting. Unfortunataley if the recurring attribute is kept, they will show up in the next meeting, which is not expected by the user. So we now remove the recurring attribute in the above scenario