Allow emails as bind names
According to RFC4511 bind names should be either null (for anonymous or SASL binds) or the string representation of a DN. But some LDAP servers appear to be working with emails instead. That's why we introduce a new setting `bindWith` that allows to define whether the users primary email address or the dn is used as the bind name.