Fix: E2E threw warning that aborted tests
Message was: 
"WARNING: the "timeoutsAsyncScript" command will be deprecated soon. If you have further questions, reach out in the WebdriverIO Gitter support channel (https://gitter.im/webdriverio/webdriverio).

Note: This command is not part of the W3C WebDriver spec and won't be supported in future versions of the driver. It is recommended to use the timeout command for this."