Fix `currentDatePlusDeltaDays`
It turned out that the old implementation of `currentDatePlusDeltaDays` didn't work exactly. The test case which passes -35 as delta in days failed since the function returned a date off by one. Ridiculously, the test case passed on all other days except today. With this small change in the implementation the tests passes again...