Fix: isOpen and isSticky attributes needed caretaking on conversion Info => Action
If optional isOpen attribute is not yet available, we must set it to true, otherwise the AI will be considered "done" immediately after conversion.
 So, a "done" state is preserved when converting back-and-forth.

As the isSticky attribute is not optional, and action items are not allowed to have isSticky attribute, we set it to false here, without condition.
So, the isSticky state is lost when conferting back-and-forth