Adapt ldap import to new MongoDB API
No idea why but mongo db now demands the db name as a parameter to client.db, even after we connected to the database and gave the db name during connect. So now we have to parse the connect url.

The mongodb client has a property s, where the db name is the value of some property of some property... but that does not look like an official API.