Fix issue with undefined topic responsibles
This can happen if the topic was created back then when we used
simple-schema. Simple-schema would set the value to null if it was
optional. Astronomy uses the default value (here: []). So new code may
assume that topic responsibles are always an array. This code does
exactly that. To prevent future errors we will create a migration so
the topic responsibles field is always an array.