Prepare topics for new minutes within meteor method to guarantee access to Topics Collection - #448
To prepare the topics for a new minutes access to the topics collection is required.
It is safer to do this preparation within a meteor methods since this will be executed
 on the server-side, too, so it is not necessary to ensure that the template executing
 the addMinutes workflow subscribed to the topics collection.

The `minutesedit.checkParent` session variable ensures that the check if the minutes is linked
 to its parent series will be done after a short timeout after the page was initialized.
So we reset it before creating new minutes.